fix(checkbox): upadte group props file
