+++ src/checkbox/type.ts
@@ -83,7 +83,7 @@ export interface TdCheckboxProps {
   };
   /**
    * 自定义选中图标和非选中图标。使用 Array 时表示：`[选中态图标，非选中态图标]`。使用 String 时，值为 circle 表示填充圆形图标、值为 line 表示描边型图标、值为 rectangle 表示填充矩形图标
-   * @default circle
+   * @default 'circle'
    */
   icon?: {
     type: null;
@@ -141,75 +141,6 @@ export interface TdCheckboxProps {
    */
   value?: {
     type: null;
-    value?: string | number;
+    value?: string | number | boolean;
   };
 }
-
-export interface TdCheckboxGroupProps {
-  /**
-   * 自定义组件样式
-   * @default ''
-   */
-  customStyle?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 是否禁用组件
-   * @default false
-   */
-  disabled?: {
-    type: BooleanConstructor;
-    value?: boolean;
-  };
-  /**
-   * 支持最多选中的数量
-   */
-  max?: {
-    type: NumberConstructor;
-    value?: number;
-  };
-  /**
-   * 统一设置内部复选框 HTML 属性
-   * @default ''
-   */
-  name?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 以配置形式设置子元素。示例1：`['北京', '上海']` ，示例2: `[{ label: '全选', checkAll: true }, { label: '上海', value: 'shanghai' }]`。checkAll 值为 true 表示当前选项为「全选选项」
-   * @default []
-   */
-  options?: {
-    type: ArrayConstructor;
-    value?: Array<CheckboxOption>;
-  };
-  /**
-   * 选中值
-   * @default []
-   */
-  value?: {
-    type: ArrayConstructor;
-    value?: CheckboxGroupValue;
-  };
-  /**
-   * 选中值，非受控属性
-   * @default []
-   */
-  defaultValue?: {
-    type: ArrayConstructor;
-    value?: CheckboxGroupValue;
-  };
-}
-
-export type CheckboxOption = string | number | CheckboxOptionObj;
-
-export interface CheckboxOptionObj {
-  label?: string;
-  value?: string | number;
-  disabled?: boolean;
-  checkAll?: true;
-}
-
-export type CheckboxGroupValue = Array<string | number>;
